Use separate header files
