+++ src/contrib/intercommunication/intercommunication_lib.h
@@ -6,25 +6,8 @@
 #ifndef ICMB_INTERCOMMUNICATION_LIB_H
 #define ICMB_INTERCOMMUNICATION_LIB_H
 
-#include <mpi.h>
-
-/*
- * For benchmarking with intra-communicators, sets the intra-communicator
- * to use.
- *
- * If no communicator is set, the benchmark will default to using
- * MPI_COMM_WORLD
- */
-void set_communicator(MPI_Comm intracommunicator);
-
-/*
- * For benchmarking with inter-communicators, sets the communicators to use for
- * the initiating and responding group, respectively.
- *
- * The inter-communicator will be constructed from these two
- * intra-communicators.
- */
-void set_communicators(MPI_Comm initiator, MPI_Comm responder);
+#include "launchers_lib.h"
+#include "utilities.h"
 
 #endif /* ICMB_INTERCOMMUNICATION_LIB_H */
