+++ src/contrib/intercommunication/launchers.h
@@ -6,7 +6,6 @@
 #ifndef ICMB_LAUNCHERS_H
 #define ICMB_LAUNCHERS_H
 
-
 void icmb_launch_client(char* port_name);
 void icmb_launch_master(int num_workers, char* command, char** argv);
 void icmb_launch_server();
