+++ src/contrib/intercommunication/options_parser.h
@@ -1,0 +1,13 @@
+/* Inter-Communicator MPI Benchmarking
+ * Copyright (c) 2021 Stefan Christians
+ * SPDX-License-Identifier: MIT
+ */
+
+#ifndef ICMB_OPTIONS_PARSER_H
+#define ICMB_OPTIONS_PARSER_H
+
+void icmb_parse_intercommunication_options (int argc, char** argv);
+void icmb_print_intercommunication_help();
+
+#endif /* ICMB_OPTIONS_PARSER_H */
+
